+++ classes/notifications_table.php
@@ -52,19 +52,19 @@ class advnotifications_notifications_table extends advnotifications_base_table {
             return get_string('advnotifications_edit_label', 'block_advnotifications') . ' | ' .
                     get_string('advnotifications_delete_label', 'block_advnotifications');
         } else {
-            return '<form id="tr'.$values->id.'" data-edit="' . $values->id . '" method="POST" action="' . $CFG->wwwroot .
+            return '<form id="tredit'.$values->id.'" data-edit="' . $values->id . '" method="POST" action="' . $CFG->wwwroot .
                 '/blocks/advnotifications/pages/process.php">
                     <input type="hidden" class="edit_notification_sesskey" name="sesskey" value="' . sesskey() . '">
                     <input type="hidden" class="edit_notification_purpose" name="purpose" value="edit">
                     <input type="hidden" class="edit_notification_tableaction" name="tableaction" value="' . $values->id . '">
-                    <input type="submit" class="edit_notification_edit btn btn-light" name="edit" value="' .
+                    <input type="submit" class="edit_notification_edit btn btn-secondary" name="edit" value="' .
                         get_string('advnotifications_edit_label', 'block_advnotifications') . '">
-                </form> <form id="tr'.$values->id.'" data-delete="' . $values->id . '" method="POST" action="' . $CFG->wwwroot .
+                </form> <form id="trdelete'.$values->id.'" data-delete="' . $values->id . '" method="POST" action="' . $CFG->wwwroot .
                 '/blocks/advnotifications/pages/process.php">
                     <input type="hidden" class="delete_notification_sesskey" name="sesskey" value="' . sesskey() . '">
                     <input type="hidden" class="delete_notification_purpose" name="purpose" value="delete">
                     <input type="hidden" class="delete_notification_tableaction" name="tableaction" value="' . $values->id . '">
-                    <input type="submit" class="delete_notification_delete btn btn-light" name="delete" value="' .
+                    <input type="submit" class="delete_notification_delete btn btn-danger" name="delete" value="' .
                 get_string('advnotifications_delete_label', 'block_advnotifications') . '">
                 </form>';
         }