+++ classes/restore_table.php
@@ -50,14 +50,14 @@ class advnotifications_restore_table extends advnotifications_base_table {
             return get_string('advnotifications_restore_label', 'block_advnotifications') . ' | ' .
                     get_string('advnotifications_delete_label', 'block_advnotifications');
         } else {
-            return '<form id="tr'.$values->id.'" data-restore="' . $values->id . '" method="POST" action="' . $CFG->wwwroot .
+            return '<form id="trrestore'.$values->id.'" data-restore="' . $values->id . '" method="POST" action="' . $CFG->wwwroot .
                 '/blocks/advnotifications/pages/process.php">
                     <input type="hidden" class="restore_notification_sesskey" name="sesskey" value="' . sesskey() . '">
                     <input type="hidden" class="restore_notification_purpose" name="purpose" value="restore">
                     <input type="hidden" class="restore_notification_tableaction" name="tableaction" value="' . $values->id . '">
                     <input type="submit" class="restore_notification_restore btn btn-light" name="edit" value="' .
                 get_string('advnotifications_restore_label', 'block_advnotifications') . '">
-                </form> <form id="tr'.$values->id.'" data-permdelete="' . $values->id . '" method="POST" action="' . $CFG->wwwroot .
+                </form> <form id="trpermdelete'.$values->id.'" data-permdelete="' . $values->id . '" method="POST" action="' . $CFG->wwwroot .
                 '/blocks/advnotifications/pages/process.php">
                     <input type="hidden" class="delete_notification_sesskey" name="sesskey" value="' . sesskey() . '">
                     <input type="hidden" class="delete_notification_purpose" name="purpose" value="permdelete">