+++ settings.php
@@ -23,6 +23,8 @@
  * @author     Zander Potgieter <zander.potgieter@learningworks.co.nz>
  */
 
+require_once('locallib.php');
+
 defined('MOODLE_INTERNAL') || die;
 
 global $CFG;
@@ -66,6 +68,18 @@ if ($ADMIN->fulltree) {
             get_string('setting/enable', 'block_advnotifications'),                                         // TITLE.
             get_string('setting/enable_desc', 'block_advnotifications'),                                    // DESCRIPTION.
             get_string('setting/enable_default', 'block_advnotifications')                                  // DEFAULT.
+        )
+    );
+
+    $options = get_date_formats();
+    // ALLOW HTML TOGGLE.
+    $settings->add(
+        new admin_setting_configselect(
+            'block_advnotifications/dateformat',                                                            // NAME.
+            get_string('setting/dateformat', 'block_advnotifications'),                                     // TITLE.
+            get_string('setting/dateformat_desc', 'block_advnotifications'),                                // DESCRIPTION.
+            key(reset($options)),                                                                           // DEFAULT.
+            $options                                                                                        // OPTIONS.
         )
     );